Fix issue #18: hicpp-signed-bitwise warnings in field arithmetic and RSA/OpenSSL code
Arithmetic right-shifts on int64_t in Solinas reduction (p256_field,
p384_field) and borrow extraction (p256_point, p384_point) are
intentional: they propagate signed carries and must remain signed.
Suppressed with NOLINT(hicpp-signed-bitwise).

uint8_t XOR/AND operands in rsa_oaep and rsa_pss integer-promote to int
before bitwise ops. Fixed by casting to uint8_t or unsigned at the
operation site.

KeyAttributes::Usage enum values integer-promote to int before |.
Fixed by casting the first operand to unsigned in all six usage-bitmask
expressions in openssl_backend.hpp.

+++ providers/arm_asm/rsa_oaep.hpp
@@ -176,7 +176,7 @@ inline bool oaep_decode(
     std::array<CryptoByte, oaep_hash_len> lhash{};
     sha384(label, label_len, lhash.data());
     for (std::size_t i = 0; i < oaep_hash_len; ++i) {
-        err |= db[i] ^ lhash[i]; // NOLINT(cppcoreguidelines-pro-bounds-constant-array-index)
+        err |= static_cast<uint8_t>(db[i] ^ lhash[i]); // NOLINT(cppcoreguidelines-pro-bounds-constant-array-index)
     }
 
     // Find the 0x01 separator in DB[hLen..db_len-1] (constant-time).
@@ -185,10 +185,10 @@ inline bool oaep_decode(
     uint8_t found = 0U;
     std::size_t msg_start = db_len;  // index into db[] of first message byte
     for (std::size_t i = oaep_hash_len; i < db_len; ++i) {
-        const uint8_t is_one  = static_cast<uint8_t>((db[i] == 0x01U) & (found == 0U)); // NOLINT(cppcoreguidelines-pro-bounds-constant-array-index)
+        const uint8_t is_one  = static_cast<uint8_t>(static_cast<unsigned>(db[i] == 0x01U) & static_cast<unsigned>(found == 0U)); // NOLINT(cppcoreguidelines-pro-bounds-constant-array-index)
         const uint8_t is_zero = static_cast<uint8_t>( db[i] == 0x00U); // NOLINT(cppcoreguidelines-pro-bounds-constant-array-index)
         // If we haven't found 0x01 yet and this byte is not 0x00 and not 0x01: error.
-        err |= static_cast<uint8_t>((found == 0U) & (is_zero == 0U) & (is_one == 0U));
+        err |= static_cast<uint8_t>(static_cast<unsigned>(found == 0U) & static_cast<unsigned>(is_zero == 0U) & static_cast<unsigned>(is_one == 0U));
         // Update msg_start in constant-time fashion.
         // When is_one transitions found 0→1, set msg_start = i+1.
         const std::size_t candidate = i + 1U;
